+++ libs/gi/ui/src/components/character/index.ts
@@ -3,7 +3,6 @@ export * from './card'
 export * from './CharacterCard'
 export * from './CharacterCardPico'
 export * from './CharacterMultiAutocomplete'
-export * from './CharacterMultiSelectionModal'
 export * from './CharacterProfilePieces'
 export * from './CharacterSelectionModal'
 export * from './CharacterSort'
